Reanimator has undergoing some major refactoring. For the time being, the main reanimator trunk (\source) has not been affected. Before Reanimator itself is refactored, the Hellgate dependency must be completed and functional.
Because we have so many different components now it is worth isolating the various classes into their own projects and namespaces. The 'Revival' solution currently contains these projects:
- Revival
|--- Common - contains common components shared by everything. FileTools etc
|--- Hellgate - contains all of the Hellgate related logic. Index, Excel, Xml, etc.
|--- Hellpack - a command line utility designed for packing .idx .dats and compiling raw excel and xml source.
|--- Reanimator - the Reanimator program.

There are other projects to add to the solution including the launcher, these will be added as they develop. This is just an inital commit. A lot has already been refactored and it does compile but its a work in progress.